feat: added root stack screen
