+++ src/navigation/AppNavigation/index.ts
@@ -1,0 +1,1 @@
+export { default as AppNavigation } from "./AppNavigation";
