+++ src/navigation/AppNavigation/styles.ts
