+++ src/navigation/page-navigators/requests.navigation.ts
@@ -1,4 +1,4 @@
-import { createPageNavigator, NavScreen } from "./utils";
+import { createPageNavigator, NavScreen } from "../utils";
 import SCREEN_NAMES from "@app/navigation/screen.names";
 import { RequestMap } from "@app/screens";
