+++ src/screens/Login/LoginScreen.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Text, View } from "react-native";
+
+const LoginScreen: React.FC = () => {
+  return (
+    <View>
+      <Text>login</Text>
+    </View>
+  );
+};
+
+export default LoginScreen;
