+++ src/screens/Login/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginScreen } from "./LoginScreen";
