+++ src/screens/index.tsx
@@ -14,3 +14,6 @@ export { default as RequestMap } from "./Request";
 
 // Volunteer screens
 export { default as Attending } from "./Volunteer";
+
+// Login screens
+export { LoginScreen } from "./Login";
